app/requirements.txt
anyio==3.6.1
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
fastapi==0.79.0
flatbuffers==2.0
h11==0.13.0
httptools==0.4.0
humanfriendly==10.0
idna==3.3
Jinja2==3.1.2
kaggle==1.5.12
MarkupSafe==2.1.1
mpmath==1.2.1
numpy==1.23.1
onnx==1.12.0
onnxruntime==1.12.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.20.1
pycodestyle==2.9.1
pydantic==1.9.1
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-slugify==6.1.2
pytz==2022.1
requests==2.28.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
sympy==1.10.1
text-unidecode==1.3
toml==0.10.2
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
watchfiles==0.16.1
websockets==10.3
